Refactoring for modules
* Move SI-related components to the `spring-integration-gcp` module
* Add `starters` sub-project
* Move auto-configurations to appropriate starters
* Reorganize dependencies a bit
* add `.jdk8` indicator files for animal-sniffer plugin
* Fix `GoogleStorageProtocolResolver` do not delegate
to the `DefaultResourceLoader` delegate, since that is already its
responsibility to deal with the `null` from its `protocolResolvers`
