+++ spring-cloud-gcp-dependencies/pom.xml
@@ -16,7 +16,7 @@
 	<description>Spring Cloud GCP Dependencies</description>
 
 	<properties>
-		<google-cloud-java.version>0.18.0-alpha</google-cloud-java.version>
+		<google-cloud-java.version>0.17.2-alpha</google-cloud-java.version>
 	</properties>
 
 	<dependencyManagement>
@@ -24,6 +24,21 @@
 			<dependency>
 				<groupId>org.springframework.cloud</groupId>
 				<artifactId>spring-cloud-gcp-core</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-gcp-pubsub</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-gcp-storage</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-integration-gcp</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
@@ -42,6 +57,25 @@
 				<artifactId>commons-logging</artifactId>
 				<version>1.2</version>
 			</dependency>
+
+			<!--Starters-->
+
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-gcp-starter-pubsub</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-gcp-starter-core</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-gcp-starter-storage</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+
 		</dependencies>
 	</dependencyManagement>
