+++ spring-cloud-gcp-pubsub/pom.xml
@@ -40,16 +40,9 @@
 			<scope>test</scope>
 		</dependency>
 
-		<!-- SI channel adapter dependencies -->
-		<dependency>
-			<groupId>org.springframework.integration</groupId>
-			<artifactId>spring-integration-core</artifactId>
-			<optional>true</optional>
-		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-gcp-core</artifactId>
-			<optional>true</optional>
 		</dependency>
 	</dependencies>
 </project>
