+++ spring-cloud-gcp-pubsub/src/test/java/org/springframework/cloud/gcp/pubsub/integration/converters/SimpleMessageConverterTest.java
@@ -9,6 +9,8 @@ import java.util.Map;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
+
+import org.springframework.cloud.gcp.pubsub.converters.SimpleMessageConverter;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageHeaders;
 import org.springframework.messaging.converter.MessageConversionException;
