+++ spring-cloud-gcp-starters/spring-cloud-gcp-starter-storage/src/main/resources/META-INF/spring.factories
