+++ spring-cloud-gcp-storage/pom.xml
@@ -8,17 +8,23 @@
 		<groupId>org.springframework.cloud</groupId>
 		<version>1.0.0.BUILD-SNAPSHOT</version>
 	</parent>
-
 	<artifactId>spring-cloud-gcp-storage</artifactId>
-
+	<description>Spring Cloud GCP Storage Module</description>
+	<properties>
+		<main.basedir>${basedir}/../..</main.basedir>
+	</properties>
 	<dependencies>
+		<dependency>
+			<groupId>org.springframework</groupId>
+			<artifactId>spring-context</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>commons-logging</groupId>
+			<artifactId>commons-logging</artifactId>
+		</dependency>
 		<dependency>
 			<groupId>com.google.cloud</groupId>
 			<artifactId>google-cloud</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-gcp-core</artifactId>
 		</dependency>
 	</dependencies>
 </project>
