+++ spring-cloud-gcp-storage/src/main/java/org/springframework/cloud/gcp/storage/GoogleStorageProtocolResolver.java
@@ -64,6 +64,6 @@ public class GoogleStorageProtocolResolver implements ProtocolResolver, Initiali
 		if (location.startsWith("gs://")) {
 			return new GoogleStorageResource(this.storage, location);
 		}
-		return this.delegate.getResource(location);
+		return null;
 	}
 }
